pyproject.toml
[project]
name = "online-store"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "docutils (>=0.22.3,<0.23.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "django-phonenumber-field (>=8.4.0,<9.0.0)",
    "phonenumbers (>=9.0.21,<10.0.0)",
    "zstandard (>=0.25.0,<0.26.0)",
    "uuid (>=1.30,<2.0)",
    "django-dump-load-utf8 (>=0.0.4,<0.0.5)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
